Some memory errors. Should fix time for quests and some problems with big instanceIDs.
(The time is readed as 64 bits in the core nearly everytime,
but always stored as 32 bits in DB,
what's the good value?
I set it to 32 bits, for now.)

+++ src/server/game/Entities/Player/SocialMgr.cpp
@@ -316,7 +316,7 @@ PlayerSocial *SocialMgr::LoadFromDB(PreparedQueryResult result, uint32 guid)
         Field* fields = result->Fetch();
 
         friend_guid = fields[0].GetUInt32();
-        flags = fields[1].GetUInt32();
+        flags = fields[1].GetUInt8();
         note = fields[2].GetString();
 
         social->m_playerSocialMap[friend_guid] = FriendInfo(flags, note);
